refactor(summary): move Summary component and summarize hook to domain folder
