+++ src/features/llm/index.ts
@@ -1,4 +1,3 @@
-export * from './SummaryViewer';
-export * from './useSummarize';
+export * from '../summary/useSummarize';
 export * from './generatePrompt';
 export * from './llmClient';
